Update and bug fix in creating synthetic networks
The script for creating synthetic networks had a bug when the network did not contain all household sizes (especially problematic when creating a small household for testing purposes). This has been fixed.
Also, the script has been changed to be more general, the maximal network size (now set to 7) can be edited in the beginning of the script and the remainder of the script adapts accordingly.
Comments are added.